+++ src/util/id_util.c
@@ -33,7 +33,7 @@ void decode_br(fbuf_t fbuf, dbuf_t *dbuf) {
     }
 
     dbuf->cc = bit_range(fbuf.ir, 9, 11);
-    dbuf->operand1 = fbuf.pc;
+    dbuf->operand1 = (int16_t)fbuf.pc;
     dbuf->operand2 = sign_extend(bit_range(fbuf.ir, 0, 8), 9);
 }
 
@@ -85,7 +85,7 @@ void decode_ld_ldi_lea(fbuf_t fbuf, dbuf_t *dbuf) {
     }
 
     dbuf->reg = bit_range(fbuf.ir, 9, 11);
-    dbuf->operand1 = fbuf.pc;
+    dbuf->operand1 = (int16_t)fbuf.pc;
     dbuf->operand2 = sign_extend(bit_range(fbuf.ir, 0, 8), 9);
 }
 
@@ -110,7 +110,7 @@ void decode_st_sti(fbuf_t fbuf, dbuf_t *dbuf) {
     dbuf->reg = vm.reg[bit_range(fbuf.ir, 9, 11)];
     pthread_mutex_unlock(&vm.reg_mutex);
 
-    dbuf->operand1 = fbuf.pc;
+    dbuf->operand1 = (int16_t)fbuf.pc;
     dbuf->operand2 = sign_extend(bit_range(fbuf.ir, 0, 8), 9);
 }
 
@@ -129,7 +129,7 @@ void decode_jsr(fbuf_t fbuf, dbuf_t *dbuf) {
         return;
     }
 
-    dbuf->operand1 = fbuf.pc;
+    dbuf->operand1 = (int16_t)fbuf.pc;
     dbuf->operand2 = sign_extend(bit_range(fbuf.ir, 0, 10), 11);
 }
